TINS-297 conversion constraints
creating operator overloads in core requires the ITypeHelper (OverloadBindings in symbols component checks upper and lower bounds) when dealing with convertible types. The ITypeHelper in return depends on implicit/explicit conversions in this case. So far it required the ICore which in turn requires the Operators. This cyclic dependency can not be resolved. Therefore, I moved the interface IConversionsProvider from core to common in order that the ITypeHelper depends on this interface instead of ICore. This way the dependency can be resolved